requirements-dev.txt
black==23.9.1
factory-boy==3.3.0
fakeredis==2.20.3
mypy==1.6.1
pre-commit==3.6.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ruff==0.1.3
